moved all java files into duke package
