+++ src/main/java/duke/Command.java
@@ -1,3 +1,5 @@
+package duke;
+
 interface Command {
 
     void execute(TaskList tasks, Ui ui) throws DukeException;
