+++ src/main/java/duke/ExitCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 class ExitCommand implements Command {
 
     public void execute(TaskList tasks, Ui ui) throws DukeException {
