+++ src/main/java/duke/Storage.java
@@ -1,10 +1,11 @@
+package duke;
+
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.Scanner;
 import java.io.IOException;
 import java.io.File;
 import java.io.FileWriter;
-
 
 class Storage {
     private File file;
